Removed working files, worked parser into todo_grammar.py
